nginx-flask-mysql/backend/requirements.txt
Werkzeug==2.2.2
Flask==2.0.2
mysql-connector==2.2.9